Heyo added Menus
Main Menu and Game Over screen are added into the game. All that's left is to add the death condition, tune the AI (and fix the weird ground clipping bug), and maybe change the spawn for the player so the game plays better. I also have to add invisible borders so the AI and player does not walk off the edge of the map.